query_debugging_scenario/requirements.txt
datasketch
networkx
tqdm
pandas
click
dotenv
pandas
sqlglot
openai
anthropic
vertexai
sqlvalidator
google-genai
retry
python-dotenv
func_timeout